Improve performance on filesystem DBI
Large corpora were very slow to get a single IGT from, because the whole thing had to be read in. Now each corpus is broken down into a single file per IGT, and they are stored in compressed XigtJSON format. This means that retrieving a full corpus is a bit slower, but its much more reasonable than the previous wait to get a single IGT.